minor edits to documentation and location of function
