+++ epam/sequences.py
@@ -4,7 +4,6 @@
 from Bio import SeqIO
 from Bio.Seq import Seq
 
-# set alphabetical order of aa for easy use across functions
 global aa_str_sorted
 aa_str_sorted = "ACDEFGHIKLMNPQRSTVWY"
